:art: Adjust Caching Tests.
