+++ test/EasyCaching.UnitTests/CachingTests/SQLiteCachingTest.cs
@@ -4,10 +4,6 @@
     using EasyCaching.SQLite;
     using Microsoft.Extensions.DependencyInjection;
     using System;
-    using Xunit;
-    using System.Threading.Tasks;
-    using System.Collections.Generic;
-    using System.Linq;
 
     public class SQLiteCachingTest : BaseCachingProviderTest
     {
@@ -35,345 +31,7 @@
             conn.Execute(ConstSQL.CREATESQL);
 
             _provider = new DefaultSQLiteCachingProvider(_dbProvider);
-            _defaultTs = TimeSpan.FromSeconds(50);
-        }
-
-        [Fact]
-        public void Get_Cached_Value_Should_Succeed()
-        {
-            var cacheKey = Guid.NewGuid().ToString();
-            var cacheValue = "value";
-
-            _provider.Set(cacheKey, cacheValue, _defaultTs);
-            var val = _provider.Get<string>(cacheKey, null, _defaultTs);
-
-            Assert.Equal(cacheValue, val.Value);
-        }
-
-        [Fact]
-        public async Task Get_Cached_Value_Async_Should_Succeed()
-        {
-            var cacheKey = Guid.NewGuid().ToString();
-            var cacheValue = "value";
-
-            await _provider.SetAsync(cacheKey, cacheValue, _defaultTs);
-            var val = await _provider.GetAsync<string>(cacheKey, null, _defaultTs);
-
-            Assert.Equal(cacheValue, val.Value);
-        }
-
-        [Fact]
-        public void Get_Cached_Value_Without_Retriever_Should_Succeed()
-        {
-            var cacheKey = Guid.NewGuid().ToString();
-            var cacheValue = "value";
-
-            _provider.Set(cacheKey, cacheValue, _defaultTs);
-            var val = _provider.Get<string>(cacheKey);
-
-            Assert.Equal(cacheValue, val.Value);
-        }
-
-        [Fact]
-        public async Task Get_Cached_Value_Without_Retriever_Async_Should_Succeed()
-        {
-            var cacheKey = Guid.NewGuid().ToString();
-            var cacheValue = "value";
-
-            await _provider.SetAsync(cacheKey, cacheValue, _defaultTs);
-            var val = await _provider.GetAsync<string>(cacheKey);
-
-            Assert.Equal(cacheValue, val.Value);
-        }
-
-        [Fact]
-        public void Get_Not_Cached_Value_Without_Retriever_Should_Succeed()
-        {
-            var cacheKey = Guid.NewGuid().ToString();
-
-            var val = _provider.Get<string>(cacheKey);
-
-            Assert.Equal(default(string), val.Value);
-        }
-
-        [Fact]
-        public async Task Get_Not_Cached_Value_Without_Retriever_Async_Should_Succeed()
-        {
-            var cacheKey = Guid.NewGuid().ToString();
-
-            var val = await _provider.GetAsync<string>(cacheKey);
-
-            Assert.Equal(default(string), val.Value);
-        }
-
-        [Fact]
-        public void Remove_Cached_Value_Should_Succeed()
-        {
-            var cacheKey = Guid.NewGuid().ToString();
-            var cacheValue = "value";
-            _provider.Set(cacheKey, cacheValue, _defaultTs);
-            var valBeforeRemove = _provider.Get<string>(cacheKey, null, _defaultTs);
-            Assert.NotNull(valBeforeRemove);
-
-            _provider.Remove(cacheKey);
-            var valAfterRemove = _provider.Get(cacheKey, () => "123", _defaultTs);
-            Assert.Equal("123", valAfterRemove.Value);
-        }
-
-        [Fact]
-        public async Task Remove_Cached_Value_Async_Should_Succeed()
-        {
-            var cacheKey = Guid.NewGuid().ToString();
-            var cacheValue = "value";
-            await _provider.SetAsync(cacheKey, cacheValue, _defaultTs);
-            var valBeforeRemove = await _provider.GetAsync<string>(cacheKey, null, _defaultTs);
-            Assert.NotNull(valBeforeRemove);
-
-            await _provider.RemoveAsync(cacheKey);
-            var valAfterRemove = await _provider.GetAsync<string>(cacheKey, async () => await Task.FromResult("123"), _defaultTs);
-            Assert.Equal("123", valAfterRemove.Value);
-        }
-
-
-        [Fact]
-        public void Refresh_Should_Succeed()
-        {
-            var cacheKey = Guid.NewGuid().ToString();
-            var cacheValue = "value";
-            _provider.Set(cacheKey, cacheValue, _defaultTs);
-
-            var tmp = _provider.Get<string>(cacheKey);
-            Assert.Equal("value", tmp.Value);
-
-            _provider.Refresh(cacheKey, "NewValue", _defaultTs);
-
-            var act = _provider.Get<string>(cacheKey);
-
-            Assert.Equal("NewValue", act.Value);
-        }
-
-        [Fact]
-        public async Task Refresh_Async_Should_Succeed()
-        {
-            var cacheKey = Guid.NewGuid().ToString();
-            var cacheValue = "value";
-            await _provider.SetAsync(cacheKey, cacheValue, _defaultTs);
-
-            var tmp = await _provider.GetAsync<string>(cacheKey);
-            Assert.Equal("value", tmp.Value);
-
-            await _provider.RefreshAsync(cacheKey, "NewValue", _defaultTs);
-
-            var act = await _provider.GetAsync<string>(cacheKey);
-
-            Assert.Equal("NewValue", act.Value);
-        }
-
-        [Fact]
-        public void RemoveByPrefix_Should_Succeed()
-        {
-            var cacheKey1 = "PREFIX:1";
-            var cacheValue1 = "value1";
-            _provider.Set(cacheKey1, cacheValue1, _defaultTs);
-
-            var cacheKey2 = "PREFIX:2";
-            var cacheValue2 = "value2";
-            _provider.Set(cacheKey2, cacheValue2, _defaultTs);
-
-            var cacheKey3 = "xxx:1";
-            var cacheValue3 = "xxx1";
-            _provider.Set(cacheKey3, cacheValue3, _defaultTs);
-
-            var tmpRes1 = _provider.Get<string>(cacheKey1);
-            var tmpRes2 = _provider.Get<string>(cacheKey2);
-            var tmpRes3 = _provider.Get<string>(cacheKey3);
-
-            Assert.Equal(cacheValue1, tmpRes1.Value);
-            Assert.Equal(cacheValue2, tmpRes2.Value);
-            Assert.Equal(cacheValue3, tmpRes3.Value);
-
-            _provider.RemoveByPrefix("PREFIX");
-
-            var res1 = _provider.Get<string>(cacheKey1);
-            var res2 = _provider.Get<string>(cacheKey2);
-            var res3 = _provider.Get<string>(cacheKey3);
-
-            Assert.False(res1.HasValue);
-            Assert.False(res2.HasValue);
-            Assert.True(res3.HasValue);
-        }
-
-        [Fact]
-        public async Task RemoveByPrefixAsync_Should_Succeed()
-        {
-            var cacheKey1 = "PREFIX:1";
-            var cacheValue1 = "value1";
-            await _provider.SetAsync(cacheKey1, cacheValue1, _defaultTs);
-
-            var cacheKey2 = "PREFIX:2";
-            var cacheValue2 = "value2";
-            await _provider.SetAsync(cacheKey2, cacheValue2, _defaultTs);
-
-            var cacheKey3 = "xxx:1";
-            var cacheValue3 = "xxx1";
-            await _provider.SetAsync(cacheKey3, cacheValue3, _defaultTs);
-
-            var tmpRes1 = await _provider.GetAsync<string>(cacheKey1);
-            var tmpRes2 = await _provider.GetAsync<string>(cacheKey2);
-            var tmpRes3 = await _provider.GetAsync<string>(cacheKey3);
-
-            Assert.Equal(cacheValue1, tmpRes1.Value);
-            Assert.Equal(cacheValue2, tmpRes2.Value);
-            Assert.Equal(cacheValue3, tmpRes3.Value);
-
-            await _provider.RemoveByPrefixAsync("PREFIX");
-
-            var res1 = await _provider.GetAsync<string>(cacheKey1);
-            var res2 = await _provider.GetAsync<string>(cacheKey2);
-            var res3 = await _provider.GetAsync<string>(cacheKey3);
-
-            Assert.False(res1.HasValue);
-            Assert.False(res2.HasValue);
-            Assert.True(res3.HasValue);
-        }
-
-        [Fact]
-        public void SetAll_Should_Succeed()
-        {
-            var dict = GetMultiDict();
-
-            _provider.SetAll(dict, _defaultTs);
-
-            var res1 = _provider.Get<string>("key:1");
-            var res2 = _provider.Get<string>("key:2");
-
-            Assert.Equal("value1", res1.Value);
-            Assert.Equal("value2", res2.Value);
-        }
-
-        [Fact]
-        public async Task SetAllAsync_Should_Succeed()
-        {
-            var dict = GetMultiDict();
-
-            await _provider.SetAllAsync(dict, _defaultTs);
-
-            var res1 = _provider.Get<string>("key:1");
-            var res2 = _provider.Get<string>("key:2");
-
-            Assert.Equal("value1", res1.Value);
-            Assert.Equal("value2", res2.Value);
-        }
-
-        [Fact]
-        public void GetAll_Should_Succeed()
-        {
-            var dict = GetMultiDict();
-
-            _provider.SetAll(dict, _defaultTs);
-
-            var res = _provider.GetAll<string>(new List<string> { "key:1", "key:2" });
-
-            Assert.Equal(2, res.Count);
-
-            Assert.True(res.Select(x => x.Key).Contains("key:1"));
-            Assert.True(res.Select(x => x.Key).Contains("key:2"));
-            Assert.Equal(res.Where(x => x.Key == "key:1").Select(x => x.Value).FirstOrDefault().Value, "value1");
-            Assert.Equal(res.Where(x => x.Key == "key:2").Select(x => x.Value).FirstOrDefault().Value, "value2");
-        }
-
-        [Fact]
-        public async Task GetAllAsync_Should_Succeed()
-        {
-            var dict = GetMultiDict();
-
-            _provider.SetAll(dict, _defaultTs);
-
-            var res = await _provider.GetAllAsync<string>(new List<string> { "key:1", "key:2" });
-
-            Assert.Equal(2, res.Count);
-
-            Assert.True(res.Select(x => x.Key).Contains("key:1"));
-            Assert.True(res.Select(x => x.Key).Contains("key:2"));
-            Assert.Equal(res.Where(x => x.Key == "key:1").Select(x => x.Value).FirstOrDefault().Value, "value1");
-            Assert.Equal(res.Where(x => x.Key == "key:2").Select(x => x.Value).FirstOrDefault().Value, "value2");
-        }
-
-        [Fact]
-        public void GetByPrefix_Should_Succeed()
-        {
-            var dict = GetMultiDict();
-
-            _provider.SetAll(dict, _defaultTs);
-
-            string prefix = "key:";
-
-            var res = _provider.GetByPrefix<string>(prefix);
-
-            Assert.Equal(2, res.Count);
-            Assert.True(res.Select(x => x.Key).Contains("key:1"));
-            Assert.True(res.Select(x => x.Key).Contains("key:2"));
-            Assert.Equal(res.Where(x => x.Key == "key:1").Select(x => x.Value).FirstOrDefault().Value, "value1");
-            Assert.Equal(res.Where(x => x.Key == "key:2").Select(x => x.Value).FirstOrDefault().Value, "value2");
-        }
-
-        [Fact]
-        public async Task GetByPrefixAsync_Should_Succeed()
-        {
-            var dict = GetMultiDict();
-
-            _provider.SetAll(dict, _defaultTs);
-
-            string prefix = "key:";
-
-            var res = await _provider.GetByPrefixAsync<string>(prefix);
-
-            Assert.Equal(2, res.Count);
-            Assert.True(res.Select(x => x.Key).Contains("key:1"));
-            Assert.True(res.Select(x => x.Key).Contains("key:2"));
-            Assert.Equal(res.Where(x => x.Key == "key:1").Select(x => x.Value).FirstOrDefault().Value, "value1");
-            Assert.Equal(res.Where(x => x.Key == "key:2").Select(x => x.Value).FirstOrDefault().Value, "value2");
-        }
-
-        [Fact]
-        public void RemoveAll_Should_Succeed()
-        {
-            var dict = GetMultiDict();
-
-            _provider.SetAll(dict, _defaultTs);
-
-            _provider.RemoveAll(new List<string> { "key:1", "key:2" });
-
-            var res1 = _provider.Get<string>("key:1");
-            var res2 = _provider.Get<string>("key:2");
-
-            Assert.False(res1.HasValue);
-            Assert.False(res2.HasValue);
-        }
-
-        [Fact]
-        public async Task RemoveAllAsync_Should_Succeed()
-        {
-            var dict = GetMultiDict();
-
-            _provider.SetAll(dict, _defaultTs);
-
-            await _provider.RemoveAllAsync(new List<string> { "key:1", "key:2" });
-
-            var res1 = _provider.Get<string>("key:1");
-            var res2 = _provider.Get<string>("key:2");
-
-            Assert.False(res1.HasValue);
-            Assert.False(res2.HasValue);
-        }
-              
-        private Dictionary<string, string> GetMultiDict()
-        {
-            return new Dictionary<string, string>()
-            {
-                {"key:1","value1"},
-                {"key:2","value2"}
-            };
+            _defaultTs = TimeSpan.FromSeconds(30);
         }
     }
 }
